backend: completed access control

+++ src/main/java/it/unical/inf/ea/trintedapp/data/service/RecensioneServiceImpl.java
@@ -1,5 +1,6 @@
 package it.unical.inf.ea.trintedapp.data.service;
 
+import it.unical.inf.ea.trintedapp.config.AppwriteConfig;
 import it.unical.inf.ea.trintedapp.data.dao.RecensioneDao;
 import it.unical.inf.ea.trintedapp.data.dao.UtenteDao;
 import it.unical.inf.ea.trintedapp.data.entities.Recensione;
@@ -13,16 +14,22 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+
+import io.appwrite.Client;
+import io.appwrite.coroutines.CoroutineCallback;
+import io.appwrite.services.Account;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class RecensioneServiceImpl implements RecensioneService{
+public class RecensioneServiceImpl implements RecensioneService {
 
     private final RecensioneDao recensioneDao;
     private final ModelMapper modelMapper;
@@ -42,7 +49,8 @@ public class RecensioneServiceImpl implements RecensioneService{
     @Override
     public RecensioneDto getById(Long id) {
         Recensione recensione = recensioneDao.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException(String.format("Non esiste una recensione con id: [%s]", id)));
+                .orElseThrow(
+                        () -> new EntityNotFoundException(String.format("Non esiste una recensione con id: [%s]", id)));
         return modelMapper.map(recensione, RecensioneDto.class);
     }
 
@@ -54,26 +62,51 @@ public class RecensioneServiceImpl implements RecensioneService{
     }
 
     @Override
-    public void delete(Long id) {
-        RecensioneDto recensione = getById(id);
-        recensioneDao.deleteById(id);
-        
-        // Update destinatario's rating accordingly
-        Optional<Utente> destinatarioOptional = utenteDao.findByCredenzialiEmail(recensione.getDestinatarioCredenzialiEmail());
-        Utente destinatario = destinatarioOptional.get();
+    public HttpStatus delete(Long id, String jwt) {
 
-        List<Recensione> recensioni = findAll(destinatario.getId());
-        if (!recensioni.isEmpty()) {
-            float sum = 0;
-            for (Recensione r : recensioni) {
-                sum += r.getRating();
-            }
-            destinatario.setRatingGenerale(sum / recensioni.size());
-        } else {
-            destinatario.setRatingGenerale(null);
+        CompletableFuture<HttpStatus> status = new CompletableFuture<>();
+
+        RecensioneDto recensione = getById(id);
+
+        // Update destinatario's rating accordingly
+        Utente destinatario = utenteDao.findByCredenzialiEmail(recensione.getDestinatarioCredenzialiEmail()).get();
+
+        Client client = new Client(AppwriteConfig.ENDPOINT)
+                .setProject(AppwriteConfig.PROJECT_ID)
+                .setJWT(jwt);
+
+        Account account = new Account(client);
+
+        try {
+            account.get(
+                    new CoroutineCallback<>((response, error) -> {
+                        Utente jwtUser = utenteDao.findByCredenzialiEmail(response.getEmail()).get();
+                        if (destinatario.getId() == jwtUser.getId() || jwtUser.getIsAdmin()) {
+                            recensioneDao.deleteById(id);
+                            List<Recensione> recensioni = findAll(destinatario.getId());
+                            if (!recensioni.isEmpty()) {
+                                float sum = 0;
+                                for (Recensione r : recensioni) {
+                                    sum += r.getRating();
+                                }
+                                destinatario.setRatingGenerale(sum / recensioni.size());
+                            } else {
+                                destinatario.setRatingGenerale(null);
+                            }
+
+                            utenteDao.save(destinatario);
+
+                            status.complete(HttpStatus.OK);
+
+                        } else {
+                            status.complete(HttpStatus.UNAUTHORIZED);
+                        }
+                    }));
+        } catch (Exception e) {
+            status.complete(HttpStatus.INTERNAL_SERVER_ERROR);
         }
 
-        utenteDao.save(destinatario);
+        return status.join();
     }
 
     private static final int SIZE_FOR_PAGE = 20;
@@ -81,50 +114,65 @@ public class RecensioneServiceImpl implements RecensioneService{
     @Override
     public Page<RecensioneDto> getAllPaged(int page) {
         Page<Recensione> recensioni = recensioneDao.findAll(PageRequest.of(page, SIZE_FOR_PAGE));
-        List<RecensioneDto> listRecensioni = recensioni.stream().map(recensione -> modelMapper.map(recensione, RecensioneDto.class))
+        List<RecensioneDto> listRecensioni = recensioni.stream()
+                .map(recensione -> modelMapper.map(recensione, RecensioneDto.class))
                 .collect(Collectors.toList());
         return new PageImpl<>(listRecensioni);
     }
 
     @Override
-    public RecensioneDto save(RecensioneDto recensioneDto) {
+    public HttpStatus save(RecensioneDto recensioneDto, String jwt) {
+        CompletableFuture<HttpStatus> status = new CompletableFuture<>();
+
+        Client client = new Client(AppwriteConfig.ENDPOINT)
+                .setProject(AppwriteConfig.PROJECT_ID)
+                .setJWT(jwt);
+
+        Account account = new Account(client);
+
         Recensione recensione = modelMapper.map(recensioneDto, Recensione.class);
 
         // Get the Utente objects based on the email
         Optional<Utente> autoreOptional = utenteDao.findByCredenzialiEmail(recensioneDto.getAutoreCredenzialiEmail());
-        Optional<Utente> destinatarioOptional = utenteDao.findByCredenzialiEmail(recensioneDto.getDestinatarioCredenzialiEmail());
+        Optional<Utente> destinatarioOptional = utenteDao
+                .findByCredenzialiEmail(recensioneDto.getDestinatarioCredenzialiEmail());
 
-        if (autoreOptional.isPresent() && destinatarioOptional.isPresent()) {
-            Utente autore = autoreOptional.get();
-            Utente destinatario = destinatarioOptional.get();
+        try {
 
-            // Set the Utente objects as the autore and destinatario
-            recensione.setAutore(autore);
-            recensione.setDestinatario(destinatario);
+            account.get(
+                    new CoroutineCallback<>((response, error) -> {
+                        if (response.getEmail().equals(destinatarioOptional.get().getCredenziali().getEmail())) {
+                            if (autoreOptional.isPresent() && destinatarioOptional.isPresent()) {
+                                Utente autore = autoreOptional.get();
+                                Utente destinatario = destinatarioOptional.get();
 
-            Recensione recensione1 = recensioneDao.save(recensione);
+                                // Set the Utente objects as the autore and destinatario
+                                recensione.setAutore(autore);
+                                recensione.setDestinatario(destinatario);
 
-            // Update rating accordingly
-            List<Recensione> recensioni = findAll(destinatario.getId());
-            if (!recensioni.isEmpty()) {
-                float sum = 0;
-                for (Recensione r : recensioni) {
-                    sum += r.getRating();
-                }
-                destinatario.setRatingGenerale(sum / recensioni.size());
-            } else {
-                destinatario.setRatingGenerale(recensione.getRating());
-            }
+                                // Update rating accordingly
+                                List<Recensione> recensioni = findAll(destinatario.getId());
+                                if (!recensioni.isEmpty()) {
+                                    float sum = 0;
+                                    for (Recensione r : recensioni) {
+                                        sum += r.getRating();
+                                    }
+                                    destinatario.setRatingGenerale(sum / recensioni.size());
+                                } else {
+                                    destinatario.setRatingGenerale(recensione.getRating());
+                                }
+                                utenteDao.save(destinatario);
+                            }
+                        } else {
+                            status.complete(HttpStatus.FORBIDDEN);
+                        }
+                    }));
 
-            utenteDao.save(destinatario);
-
-            return modelMapper.map(recensione1, RecensioneDto.class);
-        } else {
-            // Handle the case when autore or destinatario is not found
-            throw new IllegalArgumentException("Invalid autore or destinatario email");
+        } catch (Exception e) {
+            status.completeExceptionally(e);
         }
+
+        return status.join();
     }
 
-
 }
-
